+++ budgeteer/web_interface/vuejs_frontend/src/App.vue
@@ -2,11 +2,31 @@
 import "@/assets/scss/app.scss"
 import {useDark, useToggle} from "@vueuse/core"
 import {useStore} from 'vuex'
+import {useToast} from 'vue-toastification'
+import axios from "axios";
 
 const isDark = useDark()
 const toggleDark = useToggle(isDark)
 
 const store = useStore()
+const toast = useToast()
+
+function saveJSON(name) {
+  axios.post('api/json/' + name + '/', store.state.data[name])
+      .then(function () {
+        console.log(name + ' gespeichert.')
+      }, function () {
+        console.log('Konnte ' + name + ' nicht speichern!')
+        toast.error('Konnte ' + name + ' nicht speichern!')
+      })
+}
+
+function saveOnLock() {
+  saveJSON('balances')
+  saveJSON('budgets')
+  saveJSON('open_transactions')
+  store.commit('setLocked', true)
+}
 </script>
 
 <template>
@@ -21,7 +41,7 @@ const store = useStore()
     <button v-else
             class="btn btn-outline-danger"
             type="button"
-            @click="store.commit('setLocked', true)"><i class="bi bi-lock"/>
+            @click="saveOnLock()"><i class="bi bi-lock"/>
     </button>
 
     <button type="button" class="btn btn-outline-secondary bg-dark m-3 text-warning" @click="toggleDark()">