+++ .github/Changelog.md
@@ -5,5 +5,4 @@
 ---
 
 ### Changelog BudgeTeer:
-- Add automated Docker build
-- Enable version check
+- Save balances, budgets and open transactions on lock